requirements.txt
# Core ML Libraries
torch>=2.2.0  # Latest stable version (2.6.0 not yet available, use safetensors for security)
transformers>=4.30.0
accelerate>=0.26.0
sentencepiece>=0.1.99
tokenizers>=0.13.0

# FastText
fasttext>=0.9.2

# Data Processing
pandas>=2.0.0
numpy>=1.24.0,<2.0.0  # PyTorch 2.2.2 non supporta NumPy 2.x
scikit-learn>=1.3.0
datasets>=2.14.0

# MLOps
mlflow>=2.7.0
evidently>=0.4.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Monitoring & Visualization
plotly>=5.15.0
streamlit>=1.25.0
matplotlib>=3.7.0
seaborn>=0.12.0
gradio>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0

# Hugging Face
huggingface-hub>=0.16.0

# Data Validation
great-expectations>=0.18.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

